[ARIES-1383] Provide option to disable the provisioning of dependencies at install time.
FUNCTIONALITY

Don't delete the temporary file after installation with apache-aries-provision-dependencies:=resolve. The file may be needed in order to install dependencies packaged as part of
the archive. Currently temporary files will not be automatically deleted for subsystems using the new directive with a value of "resolve". They will continue to be deleted for all
other subsystems. We should look at addressing this in the future.

In StartAction, move the installation of dependencies to just before resolution so that all checks apply.

Anticipate the INSTALLING state when apache-aries-provision-dependencies:=resolve in UninstallAction.

When transitioning from INSTALLING to INSTALLED, in addition to computing and installing the dependencies, we must also compute the Provision-Resource header and update the
deployment manifest.

Make sure apache-aries-provision-dependencies directive is always present with default value when not specified.

Move directive from Subsystem-SymbolicName to Subsystem-Type header consistent with the provision-policy directive.

Refactor DirectiveFactory to use a map when creating directives.

Fix issue where subsystems could never obtain the state INSTALL_FAILED.

Fix issue where correct state transitions related to INSTALL_FAILED do not occur.

An unscoped subsystem with a value of apache-aries-provision-dependencies different than the parent should fail installation.

Do not allow the installation of a child into an INSTALLING parent.

When installing multiple subsystems as part of the same archive, the local repository should reflect the capabilities of any child subsystems so that the parent
resolution will succeed in order to create the correct sharing policy. This is necessary for the case where dependencies were packaged as part of the archive of a child
subsystem.

When starting a subsystem, child subsystems with apache-aries-provision-dependencies:=resolve and still in the INSTALLING state should have their dependencies installed first.

When locating a region to use for capability validation purposes, the search now stops at the first scoped ancestor whose sharing policy has already been set rather than
blindly using the immediate parent. This covers the case of multiple subsystems from the same archive being installed. The sharing policy is "set" upon entering the
INSTALLED state.

Fix issue where scoped children were not treated as part of the target region.

A target region is associated with a "controlling scoped subsystem". This is the subsystem without which the region would not exist. Recall that the subsystem graph is a
directed acyclic graph with one, and only one, source vertex, namely the root subsystem. A target region, therefore, consists of all successors of the controlling scoped
subsystem having no intervening scoped predecessor.

Fix issue where stopping a subsystem did not allow for the possibility of a subsystem with apache-aries-provision-dependencies:=resolve to be in the INSTALLING state.

Neither features nor implicitly installed subsystems may override the apache-aries-provision-dependencies value of a parent.

The root subsystem does not need its subsystem manifest validated.

The temporary file of an installed subsystem should not be deleted if any implicitly installed children have apache-aries-provision-dependencies:=resolve.

Cannot set the default value for apache-aries-provision-dependencies in the directive class itself since the default is based on the parent subsystem's value.

Move the computation of the default value to where the subsystem is first being created out of the raw content.

When checking the visibility of a capability from an installable resource, we want to stop walking up the region tree if the parent subsystem accepts dependencies because we know the
resource will be installed there and visible to that region.

RegionUpdater.addRequirements may now be called with a null argument. This means that the connection between a child (i.e. the tail) and the parent (i.e. the head) will be deleted. In
other words, the child no longer has an export sharing policy. This was necessary in order to support rolling back the policy if resolution fails.

Subsystems that are referenced by a starting subsystem, are in the INSTALLING state, and have apache-aries-provision-dependencies:=resolve must be resolved.

The export sharing policy of a subsystem must be enabled before attempting resolution and must be rolled back if resolution fails.

Add thread local in order to detect when the same subsystem is being started twice on the same thread. This can happen when two or more subsystems have circular dependencies.

Refactor so that the steps of the resolution process are more clear.

Installation of shared content should not be delayed when apache-aries-provision-dependencies:=resolve.

When a subsystem is being resolved or started, other subsystems sharing content must be resolved. Otherwise, the resolver hook will prevent the shared content from resolving.

Use closed version ranges when computing the Subsystem-Content header. If the Subsystem-Content header is not specified, it is assumed that the provider wants the resources
packaged in the archive, and only those resources, used. For more flexibility, the Subsystem-Content header may be specified with more open version ranges.

Starting a subsystem now requires a global lock. This is necessary in order to support apache-aries-provision-dependencies:=resolve, particularly with regard to circular dependencies.
This means that all subsystem starts will now be synchronous. This should be optimized in the future with a more granular locking strategy.

Subsystems containing a referenced bundle of a resolving subsystem as a constituent must also be resolved.

Support the ability to restart subsystems with apache-aries-provision-dependencies:=resolve in the INSTALLING state.

Add new AriesSubsystem-OriginalContent header.

When computing the current wiring for resolution, do not include bundles whose wiring is null.

Initialize the parent field of a persisted subsystem resource with the first scoped parent.

Handle non-hierarchical URIs, such as those derived from JAR URLs, as part of a persistent, INSTALLING subsystem restart.

Given apache-aries-provision-dependencies:=resolve, need to push INSTALLING parents into INSTALLED to ensure all dependencies are available for resolution.

Given apache-aries-provision-dependencies:=resolve, need to return the region of the first scoped ancestor whose sharing policy has been set for validating capabilities.

Allow for implicit subsystem installations during resolution when determining what is or is not content.

TESTING

Add test cases listed in JIRA as well as some others.

Make the SubsystemEventHandler test utility class public.

Fix SubsystemTest to handle INSTALLING apache-aries-provision-dependencies:=resolve.

Add new utility builder classes. Note that the lack of these will break several commits prior to this one.

Add the ability to build a bundle archive as bytes for use as test repository content.

Add facility for automatically stopping and uninstalling subsystems without the need for try/catch/finally blocks.

Fix tests to expect the new closed version ranges used by computed Subsystem-Content headers.